+++ src/main/java/net/geforcemods/securitycraft/network/server/SyncSSSSettingsOnServer.java
@@ -13,24 +13,36 @@ import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
 public class SyncSSSSettingsOnServer implements IMessage {
 	private BlockPos pos;
 	private DataType dataType;
+	private BlockPos posToRemove;
 
 	public SyncSSSSettingsOnServer() {}
 
 	public SyncSSSSettingsOnServer(BlockPos pos, DataType dataType) {
+		this(pos, dataType, null);
+	}
+
+	public SyncSSSSettingsOnServer(BlockPos pos, DataType dataType, BlockPos posToRemove) {
 		this.pos = pos;
 		this.dataType = dataType;
+		this.posToRemove = posToRemove;
 	}
 
 	@Override
 	public void toBytes(ByteBuf buf) {
 		buf.writeLong(pos.toLong());
 		buf.writeInt(dataType.ordinal());
+
+		if (dataType == DataType.REMOVE_POS)
+			buf.writeLong(posToRemove.toLong());
 	}
 
 	@Override
 	public void fromBytes(ByteBuf buf) {
 		pos = BlockPos.fromLong(buf.readLong());
 		dataType = DataType.values()[buf.readInt()];
+
+		if (dataType == DataType.REMOVE_POS)
+			posToRemove = BlockPos.fromLong(buf.readLong());
 	}
 
 	public static class Handler implements IMessageHandler<SyncSSSSettingsOnServer, IMessage> {
@@ -69,6 +81,8 @@ public class SyncSSSSettingsOnServer implements IMessage {
 						case CLEAR_NOTES:
 							sss.clearNotes();
 							break;
+						case REMOVE_POS:
+							sss.delink(message.posToRemove, false);
 					}
 				}
 			});
@@ -83,6 +97,7 @@ public class SyncSSSSettingsOnServer implements IMessage {
 		SOUND_OFF,
 		RECORDING_ON,
 		RECORDING_OFF,
-		CLEAR_NOTES;
+		CLEAR_NOTES,
+		REMOVE_POS;
 	}
 }