+++ Changelog.md
@@ -3,6 +3,7 @@
 - New: Items can now be dragged from JEI into the Inventory Scanner's ghost slots
 - New: Several blocks now have a new Universal Block Modifier option to disable them
 - New: The Sentry can now be deactivated by ICBM Classic's EMP blast. It can be reactivated by rightclicking it with redstone
+- New: Blocks that have been linked to a Sonic Security System can now be seen and removed in the item's and block's screen
 - Change: Reduced the Rail Mine's explosion size by 25% to make it comparable to other mines' explosions
 - API: Added DisabledOption for having a default option to disable blocks
 - API: Added Option#getKey and Option#getDescriptionKey to easily access the language keys associated with an option